refactor: a data que vem do git sai do sitemap para src/lib
O `git log` por caminho, o cache por build e o guarda `comDataUtil` moravam
dentro do `sitemap.ts`. A página do tópico precisa da MESMA resposta para o
`dateModified` do JSON-LD, e dois predicados para a mesma decisão é como o
buraco do sitemap nasceu — o sitemap escondendo a data por não confiar nela
enquanto a página a estampa seria a mesma contradição, uma volta depois.

Refatoração pura: as funções foram movidas sem uma linha de corpo alterada, e
a regra do `??` do tópico (artigo, com o `roadmap.ts` de reserva) virou
`atualizacaoDoTopico`, que era a única expressão inline que o sitemap tinha.

Provado rodando o `sitemap()` de antes e o de depois no mesmo processo, contra
o mesmo repositório: 40 URLs, 40 com `lastmod`, 21 datas distintas dos dois
lados, e a saída serializada byte a byte idêntica.
